[feat/refactor] 메인 페이지 구현
1. #1 메인 페이지를 기본 페이지로 설정
2. 디렉토리명, 파일명 규칙 정리
3. 레이아웃 및 기타 파일 정리
